style : ArtistTabPanel 분리
